@@ -8,25 +8,28 @@
   },
   "homepage": "https://arnab-datta.github.io/counter-app",
   "dependencies": {
+    "@types/node": "^16.18.126",
     "bootstrap": "^4.6.0",
     "font-awesome": "^4.7.0",
     "jquery": "^3.6.0",
     "popper.js": "^1.16.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-scripts": "2.1.8",
-    "typescript": "^3.9.10"
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": [
     ">0.2%",
@@ -35,6 +38,9 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "gh-pages": "^3.2.3"
+    "@playwright/test": "^1.55.0",
+    "@types/react": "^19.2.0",
+    "@types/react-dom": "^19.2.0",
+    "playwright": "^1.55.0"
   }
 }
